+++ sdk/servicebus/azure-servicebus/azure/servicebus/aio/management/_management_client_async.py
@@ -25,7 +25,7 @@ from ...management._generated.models import QueueDescriptionFeed, TopicDescripti
 
 from ..._common.utils import parse_conn_str
 from ..._common.constants import JWT_TOKEN_SCOPE
-from ...aio._base_handler_async import ServiceBusSharedKeyCredential
+from ...aio._base_handler_async import ServiceBusSharedKeyCredential, ServiceBusSASTokenCredential
 from ...management._generated.aio._configuration_async import ServiceBusManagementClientConfiguration
 from ...management._generated.aio._service_bus_management_client_async import ServiceBusManagementClient \
     as ServiceBusManagementClientImpl
@@ -49,12 +49,12 @@ class ServiceBusManagementClient:  #pylint:disable=too-many-public-methods
 
     :param str fully_qualified_namespace: The fully qualified host name for the Service Bus namespace.
     :param credential: To authenticate to manage the entities of the ServiceBus namespace.
-    :type credential: Union[AsyncTokenCredential, ~azure.servicebus.aio.ServiceBusSharedKeyCredential]
+    :type credential: AsyncTokenCredential
     """
 
     def __init__(
             self, fully_qualified_namespace: str,
-            credential: Union["AsyncTokenCredential", ServiceBusSharedKeyCredential],
+            credential: Union["AsyncTokenCredential"],
             **kwargs) -> None:
 
         self.fully_qualified_namespace = fully_qualified_namespace
@@ -136,10 +136,14 @@ class ServiceBusManagementClient:  #pylint:disable=too-many-public-methods
         :param str conn_str: The connection string of the Service Bus Namespace.
         :rtype: ~azure.servicebus.management.aio.ServiceBusManagementClient
         """
-        endpoint, shared_access_key_name, shared_access_key, _ = parse_conn_str(conn_str)
+        endpoint, shared_access_key_name, shared_access_key, token, token_expiry, _ = parse_conn_str(conn_str)
+        if token and token_expiry:
+            credential = ServiceBusSASTokenCredential(token, token_expiry)
+        elif shared_access_key_name and shared_access_key:
+            credential = ServiceBusSharedKeyCredential(shared_access_key_name, shared_access_key)
         if "//" in endpoint:
             endpoint = endpoint[endpoint.index("//")+2:]
-        return cls(endpoint, ServiceBusSharedKeyCredential(shared_access_key_name, shared_access_key), **kwargs)
+        return cls(endpoint, credential, **kwargs)
 
     async def get_queue(self, queue_name: str, **kwargs) -> QueueProperties:
         """Get the properties of a queue.