+++ sdk/servicebus/azure-servicebus/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 **New Features**
 * Messages can now be sent twice in succession.
+* Connection strings now support 
+* Connection strings used with `from_connection_string` methods now support using the `SharedAccessSignature` key in leiu of `sharedaccesskey` and `sharedaccesskeyname`, taking the string of the properly constructed token as value.
 
 **Breaking Changes**
 
@@ -11,6 +13,7 @@
 * Sending a message twice will no longer result in a MessageAlreadySettled exception.
 * `ServiceBusClient.close()` now closes spawned senders and receivers.
 * Attempting to initialize a sender or receiver with a different connection string entity and specified entity (e.g. `queue_name`) will result in an AuthenticationError
+* No longer export `ServiceBusSharedKeyCredential`
 
 ## 7.0.0b5 (2020-08-10)